# secp256k1: Add specialized field check for one.