**This is rebased on #2058.**

This adds a specialized function for testing if a field value is one that is constant time and faster than an equality check along with comprehensive tests for correct functionality.  As the benchmarks below show, it results in a reduction of roughly 2.5% for addition which translates to about a 1% reduction in typical signature verification time.

```
benchmark                         old ns/op     new ns/op     delta
--------------------------------------------------------------------
BenchmarkAddJacobian              580           564           -2.76%
BenchmarkScalarBaseMultJacobian   34629         33885         -2.15%
BenchmarkScalarBaseMultLarge      47468         47021         -0.94%
BenchmarkScalarMult               157832        156503        -0.84%
BenchmarkSigVerify                216043        213725        -1.07%
BenchmarkNonceRFC6979             6447          6370          -1.19%
```